backend/requirements.txt
asgiref==3.8.1
borb==2.1.25
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
cryptography==44.0.2
defusedxml==0.7.1
Django==5.2
django-cors-headers==4.7.0
django-extensions==4.1
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
dotenv==0.9.9
drf-extra-fields==3.4.0
flake8==7.2.0
fonttools==4.57.0
fpdf==1.7.2
gunicorn==23.0.0
idna==3.10
lxml==5.4.0
mccabe==0.7.0
oauthlib==3.2.2
packaging==25.0
pillow==11.2.1
psycopg2-binary==2.9.10
pycodestyle==2.13.0
pycparser==2.22
pyflakes==3.3.2
PyJWT==2.9.0
python-barcode==0.15.1
python-dotenv==1.1.0
python3-openid==3.2.0
qrcode==8.2
requests==2.32.3
requests-oauthlib==2.0.0
social-auth-app-django==5.4.3
social-auth-core==4.6.1
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
